requirements.txt
torch
torchvision
torchaudio
openai-clip==1.0.1 # This is the package for CLIP
matplotlib
Pillow
numpy